Fix t.plan bug in tests caught by tape
t.plan() was being called with the wrong number of expected assertions.
tap seems to not report this error when there are *more* assertions than
expected. When using tape, these errors are reported.

This fixes the t.plan() calls to be called with the correct number of
expected assertions for the test.